refactor: improve client-server imports by creating dedicated client interface
- Create interface.client.ts for client-safe mapping domain exports
- Update all client imports to use interface.client.ts instead of interface.ts
- Preserve server-only interface.ts for infrastructure and services
- Fix environment check in test-logger.ts to prevent client-side process access
- Update tRPC imports to use new server api types
