+++ src/app/map/Cache/Services/server-service.server.ts
@@ -1,7 +1,7 @@
 import "server-only";
 
 import { api } from "~/commons/trpc/server";
-import { CoordSystem } from "~/lib/domains/mapping/interface";
+import { CoordSystem } from "~/lib/domains/mapping/interface.client";
 import type { ServiceConfig } from "./types";
 
 /**
