+++ src/app/map/Cache/Services/server-service.ts
@@ -1,6 +1,6 @@
 import { useMemo } from "react";
 import { api } from "~/commons/trpc/react";
-import { CoordSystem } from "~/lib/domains/mapping/interface";
+import { CoordSystem } from "~/lib/domains/mapping/interface.client";
 import type {
   ServerService,
   ServiceConfig,
