+++ src/app/map/Cache/_coordinators/mutation-coordinator.ts
@@ -1,5 +1,5 @@
 import { type Dispatch } from "react";
-import { CoordSystem, type Coord } from "~/lib/domains/mapping/interface";
+import { CoordSystem, type Coord } from "~/lib/domains/mapping/interface.client";
 import type { MapItemAPIContract } from "~/server/api/types/contracts";
 import type { CacheAction } from "../State/types";
 import type { DataOperations } from "../Handlers/types";
@@ -8,7 +8,7 @@ import type { TileData } from "../../types/tile-data";
 import { cacheActions } from "../State/actions";
 import { OptimisticChangeTracker } from "./optimistic-tracker";
 import type { EventBusService } from "../../Services/EventBus/interface";
-import { MapItemType } from "~/lib/domains/mapping/interface";
+import { MapItemType } from "~/lib/domains/mapping/interface.client";
 
 export interface MutationCoordinatorConfig {
   dispatch: Dispatch<CacheAction>;
