+++ src/app/map/Canvas/Tile/Item/_coordinators/swap-preview.ts
@@ -1,6 +1,6 @@
 import type { TileData } from "../../../../types/tile-data";
 import type { TileColor } from "../../Base/BaseTileLayout";
-import { CoordSystem } from "~/lib/domains/mapping/interface";
+import { CoordSystem } from "~/lib/domains/mapping/interface.client";
 import { getColor as calculateColor } from "../../../../types/tile-data";
 
 /**
