+++ src/app/map/Canvas/Tile/Item/_utils/test-identifier.ts
@@ -1,4 +1,4 @@
-import type { Coord } from "~/lib/domains/mapping/interface";
+import type { Coord } from "~/lib/domains/mapping/interface.client";
 
 /**
  * Generates a test identifier for a tile based on its coordinates
