+++ src/app/map/__tests__/utils/mockTileData.ts
@@ -1,5 +1,5 @@
 import type { TileData } from "../../types/tile-data";
-import { CoordSystem } from "~/lib/domains/mapping/interface";
+import { CoordSystem } from "~/lib/domains/mapping/interface.client";
 
 export function createMockTileData(overrides?: Partial<TileData>): TileData {
   const defaultCoordId = "0,0";
