+++ src/app/map/_hooks/use-map-data.ts
@@ -1,6 +1,6 @@
 import { useState, useEffect } from "react";
 import { api } from "~/commons/trpc/react";
-import { CoordSystem, type Coord } from "~/lib/domains/mapping/interface";
+import { CoordSystem, type Coord } from "~/lib/domains/mapping/interface.client";
 import { adapt, type TileData } from "../types/tile-data";
 import type { URLInfo } from "../types/url-info";
 import type { MapItemAPIContract } from "~/server/api/types/contracts";
