+++ src/commons/trpc/react.tsx
@@ -8,7 +8,7 @@ import { observable } from "@trpc/server/observable";
 import { useState } from "react";
 import SuperJSON from "superjson";
 
-import { type AppRouter } from "~/server/api/root";
+import { type AppRouter } from "~/server/api/types";
 import { createQueryClient } from "./query-client";
 import { loggers } from "~/lib/debug/debug-logger";
 
@@ -130,6 +130,6 @@ export function TRPCReactProvider(props: { children: React.ReactNode }) {
 
 function getBaseUrl() {
   if (typeof window !== "undefined") return window.location.origin;
-  if (process.env.VERCEL_URL) return `https://${process.env.VERCEL_URL}`;
-  return `http://localhost:${process.env.PORT ?? 3000}`;
+  if (process.env.NEXT_PUBLIC_VERCEL_URL) return `https://${process.env.NEXT_PUBLIC_VERCEL_URL}`;
+  return `http://localhost:3000`;
 }
